logclean: Rework locking for increased testing parallelism.
This also introduces a lock into the build on the logs they are writing
to prevent logclean from touching them.

Lower HTML json lock wait time.

Add -w flag to specify how long to wait for the global logclean lock,
defaulting to not waiting as normal usage won't have this lock held.

This is after finding this on my system:
  Filesystem    Size    Used   Avail Capacity  Mounted on
  tmpfs         126G    124G    2.3G    98%    /tmp

+++ src/share/poudriere/logclean.sh
@@ -59,8 +59,9 @@ DRY_RUN=0
 DAYS=
 ALL=0
 MAX_COUNT=
+LOGCLEAN_LOCK_WAIT=0
 
-while getopts "aB:j:p:nN:vyz:" FLAG; do
+while getopts "aB:j:p:nN:vw:yz:" FLAG; do
 	case "${FLAG}" in
 		a)
 			DAYS=0
@@ -83,6 +84,11 @@ while getopts "aB:j:p:nN:vyz:" FLAG; do
 			;;
 		v)
 			VERBOSE=$((VERBOSE + 1))
+			;;
+		w)
+			# Undocumented as it is only useful for
+			# automated tests.
+			LOGCLEAN_LOCK_WAIT="${OPTARG}"
 			;;
 		y)
 			answer=yes
@@ -124,7 +130,6 @@ logclean_cleanup() {
 }
 OLDLOGS="$(mktemp -t poudriere_logclean)"
 
-slock_acquire logclean 1 || err 1 "Another logclean is busy"
 [ -d "${log_top}" ] || err 0 "No logs present"
 
 cd "${log_top:?}"
@@ -137,6 +142,8 @@ cd "${log_top:?}"
 find_broken_latest_per_pkg_links() {
 	required_env find_broken_latest_per_pkg_links PWD "${log_top:?}"
 
+	lock_have "logs_latest-per-pkg" ||
+		err 1 "find_broken_latest_per_pkg_links requires slock logs_latest-per-pkg"
 	log_links=3
 	# -Btime is to avoid racing with bulk logfile()
 	find -x latest-per-pkg -type f -Btime +1m ! -links "${log_links}"
@@ -150,6 +157,8 @@ find_broken_latest_per_pkg_links() {
 delete_broken_latest_per_pkg_old_symlinks() {
 	required_env delete_broken_latest_per_pkg_old_symlinks PWD "${log_top:?}"
 
+	lock_have "logs_latest-per-pkg" ||
+		err 1 "delete_broken_latest_per_pkg_old_symlinks requires slock logs_latest-per-pkg"
 	find -x -L latest-per-pkg -type l -exec rm -f {} +
 	# Each MASTERNAME/latest-per-pkg
 	find -x . -mindepth 2 -maxdepth 2 -name latest-per-pkg -print0 | \
@@ -161,6 +170,8 @@ delete_broken_latest_per_pkg_old_symlinks() {
 delete_empty_latest_per_pkg() {
 	required_env delete_empty_latest_per_pkg PWD "${log_top:?}"
 
+	lock_have "logs_latest-per-pkg" ||
+		err 1 "delete_empty_latest_per_pkg requires slock logs_latest-per-pkg"
 	# -Btime is to avoid racing with bulk logfile()
 	find -x latest-per-pkg -mindepth 1 -type d -Btime +1m -empty -delete
 }
@@ -182,6 +193,12 @@ elif [ ${DAYS} -eq 0 ]; then
 	reason="all builds in ${log_top} (filtered)"
 else
 	reason="builds older than ${DAYS} days in ${log_top} (filtered)"
+fi
+msg_n "Grabging logclean lock..."
+if slock_acquire -q "logclean_all" "${LOGCLEAN_LOCK_WAIT}"; then
+	echo " done"
+else
+	err 1 "Another logclean is busy"
 fi
 msg_n "Looking for ${reason}..."
 case "${MAX_COUNT:+set}" in
@@ -219,6 +236,29 @@ set)
 	;;
 esac
 echo " done"
+
+msg_n "Acquiring latest-per-pkg lock..."
+if slock_acquire -q "logs_latest-per-pkg" 0; then
+	echo " done"
+else
+	echo " skipped (locked by another process)"
+fi
+
+# Save which builds were modified for later html_json rewriting
+MASTERNAMES_TOUCHED="$(cat "${OLDLOGS:?}" | cut -d / -f 1 | sort -u)"
+msg_n "Acquiring build logs lock..."
+MASTERNAMES_LOCKED=
+for MASTERNAME in ${MASTERNAMES_TOUCHED?}; do
+	if slock_acquire -q "logs_${MASTERNAME}" 0; then
+		MASTERNAMES_LOCKED="${MASTERNAMES_LOCKED:+${MASTERNAMES_LOCKED} }${MASTERNAME}"
+		echo -n " ${MASTERNAME}"
+	else
+		echo -n " ${MASTERNAME} (skipped)"
+	fi
+done
+echo " done"
+slock_release "logclean_all"
+
 # Confirm these logs are safe to delete.
 ret=0
 do_confirm_delete "${OLDLOGS:?}" \
@@ -231,31 +271,40 @@ if [ ${ret} -eq 1 ]; then
 	logs_deleted=1
 fi
 
-# Save which builds were modified for later html_json rewriting
-MASTERNAMES_TOUCHED="$(cat "${OLDLOGS:?}" | cut -d / -f 1 | sort -u)"
-
 # Once that is done, we have a latest-per-pkg links to cleanup.
 reason="detached latest-per-pkg logfiles in ${log_top} (no filter)"
 msg_n "Looking for ${reason}..."
-{
-	find_broken_latest_per_pkg_links
-} > "${OLDLOGS:?}"
-echo " done"
-# Confirm latest-per-pkg links are OK to cleanup
-ret=0
-do_confirm_delete "${OLDLOGS:?}" \
-    "${reason}" \
-    "${answer}" "${DRY_RUN}" || ret=$?
+if lock_have "logs_latest-per-pkg"; then
+	{
+		find_broken_latest_per_pkg_links
+	} > "${OLDLOGS:?}"
+	echo " done"
+	# Confirm latest-per-pkg links are OK to cleanup
+	ret=0
+	do_confirm_delete "${OLDLOGS:?}" \
+	    "${reason}" \
+	    "${answer}" "${DRY_RUN}" || ret=$?
+else
+	echo " skipped (locked by another process)"
+fi
 
 if [ ${DRY_RUN} -eq 0 ]; then
 	msg_n "Removing broken legacy latest-per-pkg symlinks (no filter)..."
-	# Now we can cleanup dead links and empty directories.  Empty
-	# directories will take 2 passes to complete.
-	delete_broken_latest_per_pkg_old_symlinks
-	echo " done"
+	if lock_have "logs_latest-per-pkg"; then
+		# Now we can cleanup dead links and empty directories.  Empty
+		# directories will take 2 passes to complete.
+		delete_broken_latest_per_pkg_old_symlinks
+		echo " done"
+	else
+		echo " skipped (locked by another process)"
+	fi
 	msg_n "Removing empty latest-per-pkg directories (no filter)..."
-	delete_empty_latest_per_pkg
-	echo " done"
+	if lock_have "logs_latest-per-pkg"; then
+		delete_empty_latest_per_pkg
+		echo " done"
+	else
+		echo " skipped (locked by another process)"
+	fi
 else
 	msg "[Dry Run] Would remove broken legacy latest-per-pkg symlinks (no filter)..."
 	msg "[Dry Run] Would remove empty latest-per-pkg directories (no filter)..."
@@ -263,9 +312,8 @@ fi
 
 if [ ${logs_deleted} -eq 1 ]; then
 	[ "${DRY_RUN}" -eq 0 ] || err 1 "Would delete files with dry-run"
-
 	msg_n "Fixing latest symlinks..."
-	for MASTERNAME in ${MASTERNAMES_TOUCHED}; do
+	for MASTERNAME in ${MASTERNAMES_LOCKED:?}; do
 		echo -n "${MASTERNAME:?}..."
 		latest="$(find -x "${MASTERNAME:?}" -mindepth 2 -maxdepth 2 \
 		    \( -type d -name 'latest*' -prune \) -o \
@@ -281,7 +329,7 @@ if [ ${logs_deleted} -eq 1 ]; then
 	echo " done"
 
 	msg_n "Fixing latest-done symlinks..."
-	for MASTERNAME in ${MASTERNAMES_TOUCHED}; do
+	for MASTERNAME in ${MASTERNAMES_LOCKED:?}; do
 		echo -n "${MASTERNAME:?}..."
 		latest_done="$(find -x "${MASTERNAME:?}" -mindepth 2 -maxdepth 2 \
 		    \( -type d -name 'latest*' -prune \) -o \
@@ -297,7 +345,7 @@ if [ ${logs_deleted} -eq 1 ]; then
 	echo " done"
 
 	msg_n "Updating latest-per-pkg links..."
-	for MASTERNAME in ${MASTERNAMES_TOUCHED}; do
+	for MASTERNAME in ${MASTERNAMES_LOCKED:?}; do
 		echo -n " ${MASTERNAME}..."
 		find -x "${MASTERNAME:?}" -maxdepth 2 -mindepth 2 -name logs -print0 | \
 		    xargs -0 -J % find -x % -mindepth 1 -maxdepth 1 -type f | \
@@ -321,15 +369,28 @@ if [ ${logs_deleted} -eq 1 ]; then
 	echo " done"
 
 	msg_n "Removing empty build log directories..."
-	echo "${MASTERNAMES_TOUCHED}" | sed -e 's,$,/latest-per-pkg,' | \
-	    tr '\n' '\000' | \
-	    xargs -0 -J % find -x % -mindepth 0 -maxdepth 0 -empty | \
-	    sed -e 's,$,/..,' | xargs realpath | tr '\n' '\000' | \
-	    xargs -0 rm -rf
-	echo " done"
+	if lock_have "logs_latest-per-pkg"; then
+		echo "${MASTERNAMES_LOCKED:?}" | sed -e 's,$,/latest-per-pkg,' | \
+		    tr '\n' '\000' | \
+		    xargs -0 -J % find -x % -mindepth 0 -maxdepth 0 -empty | \
+		    sed -e 's,$,/..,' | xargs realpath | tr '\n' '\000' | \
+		    xargs -0 rm -rf
+		echo " done"
+	else
+		echo " skipped (locked by another process)"
+	fi
+
+	if lock_have "logs_latest-per-pkg"; then
+		slock_release "logs_latest-per-pkg"
+	fi
+
+	# unlock builds
+	for MASTERNAME in ${MASTERNAMES_LOCKED:?}; do
+		slock_release "logs_${MASTERNAME}"
+	done
 
 	msg "Rebuilding HTML JSON files..."
-	for MASTERNAME in ${MASTERNAMES_TOUCHED}; do
+	for MASTERNAME in ${MASTERNAMES_LOCKED:?}; do
 		# Was this build eliminated?
 		[ -d "${MASTERNAME:?}" ] || continue
 		msg_n "Rebuilding HTML JSON for: ${MASTERNAME}..."
